Refactor detectors, update readme with setup instructions
