+++ .env.example
@@ -24,7 +24,6 @@ YOLO_WEIGHTS_PATH="./data/detectors/yolo/yolov3.weights"
 YOLO_CONFIG_PATH="./data/detectors/yolo/yolov3.cfg"
 YOLO_CLASSES_PATH="./data/detectors/coco_classes.txt"
 YOLO_CLASSES_OF_INTEREST_PATH="./data/detectors/coco_classes_of_interest.txt"
-YOLO_DATA_PATH="./data/detectors/coco.data"
 YOLO_CONFIDENCE_THRESHOLD=0.5
 DETECTRON2_CONFIDENCE_THRESHOLD=0.5
 DETECTRON2_CONFIG_PATH="./data/detectors/detectron2/Base-RCNN-FPN.yaml"
