+++ README.md
@@ -8,8 +8,16 @@ Ivy is an open source video-based vehicle counting system which employs several 
 
 ## Setup
 - Clone this repo `git@github.com:nicholaskajoh/ivy.git`.
-- Create and/or use a virtual environment.
-- Run `pip install -r requirements.txt` to install dependencies.
+- Create and/or use a virtual environment (optional but recommended).
+- Install the dependencies in _requirements.txt_ `pip install -r requirements.txt`.
+- Choose a detector and install its dependencies where necessary (if you're not sure what to pick, we recommend you start with `yolo`).
+
+| Detector | Description | Details | Dependencies |
+|---|---|---|---|
+| `yolo` | Perform detection using models created with the YOLO (You Only Look Once) neural net. | https://pjreddie.com/darknet/yolo/ | |
+| `tfoda` | Perform detection using models created with the Tensorflow Object Detection API. | https://github.com/tensorflow/models/tree/master/research/object_detection | CPU: `pip install tensorflow-cpu` <br> GPU: `pip install tensorflow-gpu` |
+| `detectron2` | Perform detection using models created with FAIR's Detectron2 framework. | https://github.com/facebookresearch/detectron2 | `python -m pip install 'git+https://github.com/facebookresearch/detectron2.git'` (https://github.com/facebookresearch/detectron2/blob/master/INSTALL.md) |
+| `haarcascade` | Perform detection using Haar feature-based cascade classifiers. | https://docs.opencv.org/3.4/db/d28/tutorial_cascade_classifier.html | |
 
 ## Run
 - Create a _.env_ file (based on _.env.example_) in the project's root directory and edit as appropriate.
