+++ detectors/detector.py
@@ -2,6 +2,9 @@
 Detectors entry point.
 '''
 
+# pylint: disable=import-outside-toplevel
+
+
 def get_bounding_boxes(frame, model):
     '''
     Run object detection algorithm and return a list of bounding boxes and other metadata.
@@ -14,9 +17,7 @@ def get_bounding_boxes(frame, model):
         from detectors.tfoda import get_bounding_boxes as gbb
     elif model == 'detectron2':
         from detectors.detectron2 import get_bounding_boxes as gbb
-    elif model == 'tfoda_new':
-        from detectors.tfoda_new import get_bounding_boxes as gbb
     else:
-        raise Exception('Invalid detector model, algorithm or API specified (options: yolo, tfoda, tfoda_new, detectron2, haarcascade)')
+        raise Exception('Invalid detector model, algorithm or API specified (options: yolo, tfoda, detectron2, haarcascade)')
 
     return gbb(frame)
